+++ packages/plugin-collapsible-list/web/src/consts.ts
@@ -1,0 +1,3 @@
+export const collapsibleModals = {
+  settings: 'collapsibleSettings',
+};