+++ packages/ricos-converters/web/src/tiptap-converters.ts
@@ -77,7 +77,10 @@ export const toTiptap = (content: RichContent): JSONContent => {
 };
 
 export const toTiptapNodeAttrs = (node: Node): TiptapNode['attrs'] =>
-  new TiptapNodeBidiTransfoms(nodeConverters).toTiptap().byType(node).convert(node).attrs;
+  new TiptapNodeBidiTransfoms(nodeConverters)
+    .toTiptap()
+    .byType(node)
+    .convert(node, ricosNodeVisitor).attrs;
 
 export const fromTiptap = (content: JSONContent): RichContent => {
   return {